// gallery.enums.ts
export enum ZOOM_ACTIONS_ENUM {
  ZOOM_IN = 'zoomIn',
  ZOOM_OUT = 'zoomOut',
  ZOOM_CLEAR = 'zoomClear',
  ZOOM_MAX = 'zoomMax',
}

export enum ZOOM_STATE {
  ZOOM_IN = 'zoomIn',
  ZOOM_OUT = 'zoomOut',
}

export enum EVENTS_ERRORS {
  DOWNLOAD = 'Не удалось скачать. Попробуйте позже',
  DELETE = 'Не удалось удалить. Попробуйте позже',
  OPEN_PDF = 'Не удалось открыть PDF. Попробуйте позже',
}